implement image component [wip]
